Change detection routine to count marked pixels
While trialing with an actual exam, numbers in large wide boxes were
missed. Moving from cumulative luminosity (which is ambiguous about
number of marked pixels and their colour) to counting how many pixels
cross a difference threshold (1%), and then result is true/marked if
more than 2% of pixels are marked. This works well for square boxes but
is currently failing to detect the number one in wide boxes looking for
question numbers in real exams and needs further adjustment.