Add aspect-ratio-sensitive optical box threshold
Wide check boxes with "low-area" numbers (like the number 1) are not
being detected by the old optical box routine which was originally
intended only for check marks in square boxes. Adjusting the threshold
for all boxes does not work, because the required threshold in the wide
box is so low that it can be reached in other boxes if some "imperfect"
pixels from compression are near the border. In this mod, the threshold
of two percent is scaled by the aspect ratio, so that it reduces enough
to trigger on a single character. This also makes long boxes more
susceptible to border leakage, so this must be handled by increasing the
boundary shrinkage (currently 4 pixels on each side). Tests now also
passing on a single page filled in with a stylus in parsesvg/coord.