Explicitly identify textfields in custom pagedata
Merging pages from flattened processed papers is based on logic such as
whether a file has been seen, marked etc, and this can only be judged
from the custom textfields that are loaded into pagedata when the
flattened. Since textfields can have any name, there is no way to tell
another custom Field in the pagedata from something read in from a
textfield. This could potentially lead to a page that has custom data,
but has not been seen by a marker (for example), being treated as if it
has in fact been marked, because of the presence of a non-blank Field in
the pagedata. Ultimately this would only result in a delay, because any
page with work done on it will be included in any subsequent merged
files. However, it would be puzzlingly blank, and labelled as being
marked. By the time this is noticed, the original marker might have
quite a context switch to come back to marking the unseen page. So we
can head all that off by adding a prefix to identify textfields. This is
a string prefix "-tf". Note that we retain the "-optical" as a suffix so
that a textfield, read optically, is saved with the key
"tf-<original-name>-op[tical", whereas the value read from the textfield
extraction would be the shorter "tf-<original-name>".