Add summarisePage to help with merging
When we receive back work from multiple staff, e.g. marked scripts from
a few different markers, we have to extract out the work they have done,
either by keying into textfields, or marking with a stylus (which we do
in flattenProcessed). Any given page might be bad, or be seen, but have no
marks, or have marks, or not have even been seen at all (skipped). We need to
figure out what category a particular page is (bad, ok, skipped, seen)
so that we can decide whether to put into the merged document, as well
as extract a few key details (original document, original pageNumber, current
path to the doc we'd want to put into the merged version, who did the
marking/moderating/checking etc on this document. This function does
that. It doesn't organise or map the pages - but it provides the
necessary data for doing that.