Add createPageList to merge
Figures out which pages have unique work and need to be included in the
merged document, whilst ensuring if a page is not containing work that a
copy of it is included. Looks at all potential merge pages to make a
custom summary message for each, which is intended to be used as prefill
text to give context.