Enable WrapWidth and Angle for Paragraphs
TextPrefills have a number of styling options, including the rotation
angle of the text, and the wrap-width (with wrap enabled by
default). The option to control wrap enabling is not implemented at
present because the struct has it phrased as EnableWrap, so all the
existing code that does not set it, would break (well there might not be
much of it so far), because the default is on. So that needs changing to
DisableWrap in a potential future update. Meanwhile, make the WrapWidth
large to achieve the same effect.