Add initial attempt at mergeOverlay
MergeOverlay needs to operate just differently enough to Overlay that it
needs writing in a separate function, to avoid changing a trusted
function already in use in production. MergeOverlay is relatively
simple, because it uses the image generated in the previous flatten
step, copies the pagedata across (no change needed) and adds a merge
sidebar. It does not need to flatten, read textfields, optically or from
the pdf, or manipulate pagedata (probably don't need to update pagedata
twice just because one logical process - get ready for moderation - is
in two separate implementation steps - flatten then merge). Building
passing, but calling function not finished, and no tests yet.