Fix false positive optical box detection
Shrink the optical box bound by a futher 2 pixels on each side, bringing
it to -6 on each side. For a 175 dpi image, a 10mm x 10mm box has 69x69
pixels so this reduction takes the analysed image to 57x57 which is
still large enough to provide a valuable indication that the box has
been marked. Why shrink it further? The first two marker sets came back
and on one set, every "bad-page" box was labelled as mark-detected in the
optical check. The route cause has not been established beyond all doubt
but this bound shrinkage caused the false positives to stop, so it is
seems likely that the affected set of scripts were altered in such a way as
to cause non-white pixels near the box boundaries. This could happen if
the image is resampled to a lower quality, but the 1% difference in file
size does not seem large enough to be responsible for this. The other
possible cause for this - for those from the future - is updating the
template between when the exam was produced and when the exam is read -
which rather suggests that coordinates need to be stored in the
pagedata at time of creation. Since we are part way through the diet
now, we won't do that. Anyway, template changes can and should be
explicitly tracked rather to avoid the issue.