Add flattening to mergeOverlay to capture comments
The previous approach forgot that comments are text over the top of the
image, so we can't just use the image made during the flattening stage
because it is missing the comments. And we need to keep the comments, so
we have to run a second flattening stage. There's a tempting premature
optimisation here, but actually, rendering pages is a lot cheaper than
messing something up trying to be clever with a single flattening
stage that also does the merge, especially if multiple markers are
involved. And if future plans include NOT sending own work back to
colleagues (not currently the plan, but you never know).