Fix pagedata linked list behaviour
This fix makes sure that the UUIDs are updating properly in overlay and
mergeOverlay so we can track a page's history. We do this through the
linked list of UUID that are assigned at each stage of a page's
processing. For example, PageDetail.Current.Follows holds the UUID
of the previous stage (which can be found in the list of previous
PageDetail). The first stage in the processing has a blank, by
convention to say that it follows no other UUID.
The same linked list should be possible by using the UUID of the
PageDetail.Current.Own.UUID and PageDetail.Current.Original.UUID,
although current test files have early entries in Own/Original that did
not update correctly, so the linked list behaviour is only clear in the
latter processes. So don't take the pdf example in expected/visual to be a
golden example of what should happen -> it should be a linked list for
both cases (and in future we should replace this example just to ensure
this error is not treated as something actually want). These are
probably  redundant, but the own/original UUID were originally conceived
as risk management alternative to using long/difficult paths as keys in
maps. The error in the earlier processing is well understood and hence
does not invalidate the test (and redoing the pretend marking would be
too much faff for this fix)