Keep long comments but display them short
Some markers are pre-twitter era in the length of writing in their
sticky notes, which is great from a feedback point of view, but my
comment flattening layout tool can't handle multi-line comments, so I am
truncating them to 120 chars and adding a three-dot continuation
indicator. The full length of the comments is retained in the pagedata,
for the record, as it were.