Refactor paths for consistency
The number of custom path functions has been reduced, by passing the
desired dir as an argument, rather than baking it into the function
name. This change required changes across ca 20 files. However, all
tests are passing again. And it should be easier to add in some upcoming
features that depend on the paths as a result. The paths have been
rationalised too, attempting not to break backward compatibility
although this has not been tested.