Make optical flags global for resuse
We can't define the same flag more than once so do we have to live with
global flags that only apply to some commands? Presumably there is a way
around that, which avoids having to have all non-users understand
external concerns and alert when non-supported flag is used? It could be
done with a common function, but, sigh, it just seems wrong. The
conflict is that both Enter and Flatten  read the optical boxes
to see whether stylus marks have been made, and it seems non-obvious
that Enter needs to read anything if the results have just come from
processing by marked procesed .... to be continued.