Change moderation bar to ok/bad options only
Reduce design to just two boxes to tick (plus marks and comments) so
there are only four possible outcomes to handle (plus any mark
adjustments and comments).The states are:

!ok !bad = skipped
!ok bad = bad
ok !bad = ok
ok bad = bad (probably some human error here, but we play safe and flag it)

The original design had a third "not sure" option, doubling the number
of states to eight - and there are not eight useful states that a page
can be in after moderation. It is either ok to carry on with no further
action, orsome action needs to be taken (whether because it was
inadverently skipped or there is an issue). To flag a "not sure", the
moderator ticks "issue", then shares file read only with marker. If the
discussion leads to agreement that there is an issue, the file is
submitted as is with the rest of the batch. If there is agreement that
there is no actual issue, then the comments are left in place, but the
agreed box is ticked, making the page "ok." The query/issue box should be
unticked, or else the page will stay "bad."