Add textfield coordinate extraction
We need the textfield coordinates so we can crop the flattend page image
and search for possible stylus marks in the flattened image. Til now we
used the "known" template to give the coordinates, but this requires
keeping the same template on hand, and means that the process is brittle
to changes in templates / needs to explicitly know about layouts
etc. Whereas, that information is stored in the PDF document, and now with this
modification we can get it out.