Update extract to return page size
When we find the optical value of the text fields we need to know the
page size, so we get it from the MediaBox on each page. It's not trivial
to link directly to the page from the textfield (itself not a single
object) and even unipdf leaves one of the page-from-object methods
unimplemented. The "docN" object which is parent to the textfields on
that page seems to have the page as a parent, but accessing the mediaBox
from that object was one hurdle, and the other was that it meant tracing
that object to its kids. It seemed more pragmatic just to use an already
existing exported function to get the Media box from the pdfReader that
we had already had to use anyway.