Fix path management for moderation and enter
After marker-processed, the files are used for moderate and enter. Each
has two categories, active and inactive. The file movement was updated
so that

1. files are not removed from marker-processed (we don't know which
order the user would like to prepare moderation and enter bars, and we
should not force them to do them at the same time, because they are
different processes producing files going to different people

2. files are now copy.CopyIfNewerThanDestinationInDir, with logging of
does and does not get copied, and why (not new, error); this code was
modified from move.MoveIfNewerThanDestinationInDir, and relies on a new
"force" argument added to copy.Copy

Has been checked on real script set (although one with only keyboard
marked comments)