Fix migrate command and add test flag
The default is now test mode, which displays the planned actions. This
is partly a hedge against some sort of weird non-empty directory-eating
behaviour on some little-known OS. Obviously that should not be possible
given the guarantees of of os.Remove to only work on empty
directories. Then if the displayed actions look ok, the user can reissue
the command with the test flag set to false, and it will implement, only
printing to console if there is an error. If you run migrate in test
more again, there should be no proposed actions - if the migrate was successful.