Fix exporting and add to TestAddBars
Exports were refactored to reduce boilerplate, and added to TestAddBars
so that the whole process can be tested. The exported files were "aged"
by a new function MoveExportedFilesToIngest which handles both shifting
a specific set of files from export to ingest, but also updating the
modtime to the current time so that the file doesn't come back to
<stage>Ready as being unmarked (even though, of course, it is
unmarked).