Add linked list test for pagedata
This test maps the links in the current and previous page data on a
page, checking that the UUID in current.UUID and current.Follows can
make a linked list, the same length as the number of the
pagedatas (len(previous)+1 to include current). The order of the list in
pagedata.Previous does not matter. This test is for helping check the
lineage of processed papers for audit purposes