Fix over-sensitive optical detection
During TestAddBars, skipped checked scripts were sometimes being given
the status of marked despite being blank. The default shrinkage was
upped to 10 pixels after checking in parsesvg/coord cropped images that
this was not going to trim out a narrow left-hand-aligned "1" at left
hand side of a box, and give a false negative. But there is plenty left
at the quality we are dealing with, and could increase more if false
positives crop up during operation