Standardise pagedata at all merge steps
Pagedata now
- includes a copy of merge message if merged file (putting in as struct too much
refactoring for just now)
- old page data is removed*
- all merge stages have a check in TestAddBars that merge-message is
present, and that exactly one of each of "tf-page-ok", "tf-page-bad",
"tf-page-ok-optical" and "tf-page-bad-optical" are present. This check
helpfully caught lingering non-standard naming in the moderate sidebar.

*Caveat: mergeOverlay continues to "roll-over" pagedata because
flatten-merge is a two step process and the data that is collected in
the flatten stage needs to be in the current pagedat in the merged
file. But, the normal overlay (used in flatten) does not roll over data
so as to avoid getting optical fields from each stage mixed up- they
duplicate each other otherwise and would have required choosing based on
ordering in pagedata text, which could be fragile and adds work. So we
skip that