Fix mergeReport cross-referencing
The cross-ref throwing false positive results due to bad logic in the
mapping (something was on the wrong side of a bracket). That's been
fixed, and made a bit cleverer to boot, explicitly declaring that it is
comparing the unqiue pages in each set. The console report formatting has been tidied up, and manual testing with removal of files
from 05-anonymous triggers MISSING-ORIGINAL-PAGE-FOR error, while
removal of files from 24-marker-processed triggers
MISSING-PROCESSED-PAGE-FOR. An automatic csv version of the report (just
the line by line) is put in 99-reports to, named with the exam, process
stage and unix time (so that you can autoreport without overwriting
previous ones, if you are tracking an issue)