Add CSV library for writing marks
Maps and key-sorting are used to produce dynamic column layouts.
The header order can be specified for columns on the left hand side.
Headers can be required, that must be displayed (but in natural order)
Empty values take "-" to avoid messing with averages in reports
Testing is via writing to string, and io.Writer interface is provided