Add revision field to pagedata
We need a way to be able to edit pagedata in a file, to correct for
errors by actors or to implement office-decisions on issues. Pagedata is
hard to erase, so we need a way to update it without needing to delete
text from earlier revisions of it. We already use revision numbers for
learn receipts in another library, so it seems idiomatic for gradeX to
do the same here. The pagedata with the highest revision is reported
back from UnMarshallAllFromFile(). Test passing.