Add AddPageDataToPDF
It may be useful to be able to update the pagedata in a PDF file without
performing any other editing on it, whilst retaining any forms
data. This achieves that. Test passing. If there are hand written
annotations in the file, from a stylus, then you will almost certainly
lose them with this method. Better in that case to use ingester's
flatten <stage> <exam> --ensure-ancestor
In that approach, the page is flattened, and new page data is written,
without copying across the old page data.