Add mark and question substitution files
Since it is not possible to automatically figure out all possible human
errors in filling in the mark boxes, we need a way to have a human
define what substitutions should be made. We do it in two steps - first
we assign marks to particular question labels associated with a
file (a typical error is to put the mark value in the number
field). Then we specify what question that label should belong to. The
mark we assigned in the first mark-substitution file is then added to
any mark value already accumulated by properly named labels in that file.