Add front cover and test rules
Adds the final front cover design and the logic to respond to checker
substitution rules in the fix boxes. The final cover does not use
file-based rules, it only uses what is in the fix box, and only on that
paper. This is to help with auditing, by making only predicatable changes