Fix flatten checked textfield conflict
Merging the cover page with an already merged document resulted to two
sets of textfields being assigned to page 1, and an offset by one of
textfield and opticl field data (yielding twice the number of bad page
alerts) because the regexp that
works out what page the key is from, did not take into account the docNN
namespacing (which is not present on the first page, then starts at doc2
for the second page). An explicit check for doc is now included.