Fix old comments detected in check report
Comment counts are now tracked to allow new comments to be stacked on
top of old ones, without having to retain the actual comments
themselves. This allows the comments to be optionally converted from
actual comments to a count by setting OmitPreviousComments to true in
the overlay command or overlay task struct. We do this in the
flattenprocessed stage, where any old comments are from a previous
stage. This avoids old comments being picked up and listed as QBOX lines
in the checked reports.